ci: work around temporary Azure Pipelines issue
Our test suite is not quite well-prepared to cope with ACLs; on Ubuntu
16.04, the `chmod -w` call in the `SANITY` prereq simply won't work if
there are *any* ACLs stored in the POSIX extended attributes (that are
supported by the ext4 file system used in Azure Pipelines' Linux
agents). While this seems to be an issue only with version
8.25-2ubuntu3~16.04 of the `coreutils` package, not with version
8.28-1ubuntu1 that is shipped with Ubuntu 18.04, we currently only have
access to build agents running Ubuntu 16.04.

A fix (and an incidental workaround) are in the works, let's carry this
patch for a couple of days until it is no longer necessary.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>